Address experiment mode PR review
