+++ app/tests/unit/test_experiment_service_guardrails.py
@@ -1,8 +1,12 @@
 from __future__ import annotations
 
+from contextlib import contextmanager
 from datetime import UTC, datetime
 
+from psycopg2.extras import Json
+
 from app.core.prompts import EXPERIMENT_AGENT_SCOPE_REFUSAL
+from app.services.data.managers.experiment_manager import ExperimentManager
 from app.services.experiment_service import ExperimentService
 
 
@@ -145,6 +149,31 @@ class FakeExperimentManager:
         return [dict(self.thread)]
 
 
+class FakeCursor:
+    def __init__(self, *, fetchone_results=None):
+        self._fetchone_results = list(fetchone_results or [])
+        self.executed = []
+
+    def execute(self, query, params=None):
+        self.executed.append((query, params))
+
+    def fetchone(self):
+        if not self._fetchone_results:
+            return None
+        return self._fetchone_results.pop(0)
+
+
+def _patch_db_context(monkeypatch, manager, cursor):
+    def fake_get_db_context():
+        @contextmanager
+        def _ctx():
+            yield None, cursor
+
+        return _ctx()
+
+    monkeypatch.setattr(manager, "get_db_context", fake_get_db_context)
+
+
 def test_send_user_message_blocks_non_recipe_prompt_without_llm_call() -> None:
     text_call_count = 0
 
@@ -262,3 +291,36 @@ def test_build_context_payload_includes_full_recipe_content() -> None:
             ],
         }
     ]
+
+
+def test_create_thread_insert_omits_mode_column(monkeypatch) -> None:
+    manager = ExperimentManager()
+    cursor = FakeCursor(
+        fetchone_results=[
+            {
+                "id": "thread-1",
+                "title": "Weeknight curry",
+                "metadata": {"orchestration": "langgraph-ready"},
+                "created_by_user_id": "user-123",
+                "created_at": None,
+                "updated_at": None,
+            }
+        ]
+    )
+    _patch_db_context(monkeypatch, manager, cursor)
+
+    thread = manager.create_thread(
+        title="Weeknight curry",
+        metadata={"orchestration": "langgraph-ready"},
+        created_by_user_id="user-123",
+    )
+
+    assert thread["id"] == "thread-1"
+    query, params = cursor.executed[0]
+    assert "INSERT INTO experiment_threads" in query
+    assert "mode" not in query
+    assert len(params) == 3
+    assert params[0] == "Weeknight curry"
+    assert isinstance(params[1], Json)
+    assert params[1].adapted == {"orchestration": "langgraph-ready"}
+    assert params[2] == "user-123"
